requirements.txt
-e git://github.com/callowayproject/django-news-sitemaps.git@55da6b72580dd6d79380f2bb17240e7afa00555a#egg=django_news_sitemaps-master
-e git://github.com/dhan88/django-mailgun.git@df4e2f444bd488b6234f87c879a8402e7c8b552e#egg=django_mailgun-master
-e git://github.com/jezdez/celery-haystack.git@6d6ef5ea61a2a88bda3e65be69115841545ac3c3#egg=celery_haystack-develop
BeautifulSoup==3.2.1
Django==1.8.6
Fabric==1.10.2
Markdown==2.3.1
Pillow==2.8.2
Unidecode==0.4.14
Whoosh==2.6.0
amqp==1.4.5
anyjson==0.3.3
astroid==1.3.6
billiard==3.3.0.18
bleach==1.4.1
boto==2.27.0
celery==3.1.10
coverage==3.7.1
cssselect==0.7.1
dill==0.2.4
django-activity-stream==0.6.0
django-adaptors==0.2.3
django-appconf==0.6
django-autocomplete-light==2.2.3
django-celery-transactions==0.3.1
django-celery==3.1.17
django-compressor==1.5
django-dbsettings==0.8.2
django-debug-toolbar==1.3.0
django-epiceditor==0.2.1.2
django-filer==0.9.11
django-filter==0.10.0
django-haystack==2.4.0
django-keyedcache==1.5.1
django-localflavor==1.1
django-lot==0.0.6
django-markdown-deux==1.0.4
django-mathfilters==0.3.0
django-mptt==0.7.4
django-nose==1.4.2
django-polymorphic==0.7.1
django-pylibmc-sasl==0.2.4
django-rest-swagger==0.3.2
django-reversion==1.8.7
django-robots==1.1
django-secure==1.0
django-sslify==0.2
django-storages==1.1.8
django-tablib==3.1.1
django-taggit==0.14.0
django-watson==1.1.8
django-xmlrpc==0.1.3
djangorestframework==3.1.3
dogslow==0.9.7
easy-thumbnails==2.2
factory-boy==2.6
gdata==2.0.17
gevent==0.13.7
greenlet==0.3.4
gunicorn==0.14.3
ipython==0.10
iron-cache==0.3.2
iron-celery==0.3.1
iron-core==1.1.1
iron-mq==0.7
kombu==3.0.14
logilab-common==0.63.2
lxml==3.1.0
markdown2==2.2.1
mock==0.7.0b2
newrelic==2.18.1.15
nose==1.3.7
numpy==1.9.1
pascha==0.5
psycopg2==2.6.1
pylibmc==1.2.3
pylint-django==0.6.1
pylint-plugin-utils==0.2.2
pylint==1.4.3
pysolr==3.0.5
python-dateutil==2.2
python-memcached==1.48
pytz==2014.4
raven==5.0.0
redis==2.4.10
requests==2.3.0
scikit-learn==0.16.1
scipy==0.14.0
six==1.7.3
sqlparse==0.1.15
stripe==1.27.1
tablib==0.10.0
waitress==0.8.8
webassets==0.7.1
